Add usePopupAnalytics composable and refactor popups
